convert to a library
